[SW-11367] Add header_only, with_host_headers, with_device_headers options for...

See merge request software/dnn-library!184

+++ dnn-library/conanfile.py
@@ -1,9 +1,10 @@
 from conan import ConanFile
-from conan.tools.cmake import CMakeToolchain, CMake
+from conan.tools.cmake import CMakeToolchain, CMakeDeps, CMake
 from conan.tools.layout import cmake_layout
 from conans import tools
 from conans.errors import ConanInvalidConfiguration
 import os
+import shutil
 
 
 class DnnLibraryConan(ConanFile):
@@ -18,11 +19,17 @@ class DnnLibraryConan(ConanFile):
     settings = "os", "compiler", "build_type", "arch"
     options = {
         "fPIC": [True, False],
-        "warnings_as_errors": [True, False]
+        "warnings_as_errors": [True, False],
+        "header_only": [True, False],
+        "with_device_headers": [True, False],
+        "with_host_headers": [True, False],
     }
     default_options = {
         "fPIC": True,
-        "warnings_as_errors": False
+        "warnings_as_errors": False,
+        "header_only": True,
+        "with_device_headers": False,
+        "with_host_headers": True
     }
 
     scm = {
@@ -30,7 +37,6 @@ class DnnLibraryConan(ConanFile):
         "url": "git@gitlab.esperanto.ai:software/dnn-library.git",
         "revision": "auto",
     }
-    generators = "CMakeDeps"
 
     python_requires = "conan-common/[>=0.5.0 <1.0.0]"
 
@@ -41,10 +47,26 @@ class DnnLibraryConan(ConanFile):
         if self.settings.os == "Windows":
             del self.options.fPIC
     
-    def requirements(self):
-        self.requires("et-common-libs/0.0.5")
+    def configure(self):
+        if self.options.header_only:
+            del self.settings.arch
+            del self.settings.build_type
+            del self.settings.compiler
+            del self.settings.os
 
-    def validate(self):        
+    def requirements(self):
+        if self.options.with_device_headers:
+            self.requires("et-common-libs/[>=0.0.5 <1.0.0]")
+
+
+    def package_id(self):
+        if self.options.header_only:
+            self.info.header_only()
+    
+    def validate(self):
+        if self.options.header_only:
+            return
+        
         if self.settings.arch != "rv64":
             raise ConanInvalidConfiguration("Cross-compiling to arch {} is not supported".format(self.settings.arch))
 
@@ -58,25 +80,57 @@ class DnnLibraryConan(ConanFile):
                 raise ConanInvalidConfiguration("{0} requires {1} package with '-o {1}:{2}'".format(self.name, "et-common-libs", flag))
 
     def layout(self):
-        cmake_layout(self)
-        self.folders.source = "."
+        if not self.options.header_only:
+            cmake_layout(self)
+            self.folders.source = "."
     
     def generate(self):
-        tc = CMakeToolchain(self)
-        tc.variables["CMAKE_INSTALL_LIBDIR"] = "lib"
-        tc.variables["ENABLE_WARNINGS_AS_ERRORS"] = self.options.warnings_as_errors
-        tc.generate()
+        if not self.options.header_only:
+            deps = CMakeDeps(self)
+            deps.generate()
+
+            tc = CMakeToolchain(self)
+            tc.variables["CMAKE_INSTALL_LIBDIR"] = "lib"
+            tc.variables["ENABLE_WARNINGS_AS_ERRORS"] = self.options.warnings_as_errors
+            tc.generate()
     
     def build(self):
-        cmake = CMake(self)
-        cmake.configure()
-        cmake.build()
+        if not self.options.header_only:
+            cmake = CMake(self)
+            cmake.configure()
+            cmake.build()
 
     def package(self):
-        cmake = CMake(self)
-        cmake.install()
-        tools.rmdir(os.path.join(self.package_folder, "lib", "cmake"))
+        if self.options.header_only:
+            host_headers_dst = os.path.join("include", "dnn_lib", "host_headers", "dnn_lib")
+
+            if self.options.with_host_headers and not self.options.with_device_headers:
+                self.copy("InstrTableGenerated.h", src=os.path.join(self.source_folder, "include", "host_headers"), dst=host_headers_dst)
+            else:
+                self.copy("*", src=os.path.join(self.source_folder, "include"), dst=os.path.join("include", "dnn_lib"))
+                src = os.path.join(self.package_folder, "include", "dnn_lib", "host_headers", "InstrTableGenerated.h")
+                dst = os.path.join(self.package_folder, "include", "dnn_lib", "host_headers", "dnn_lib", "InstrTableGenerated.h")
+                os.mkdir(os.path.join(self.package_folder, "include", "dnn_lib", "host_headers", "dnn_lib"))
+                shutil.move(src, dst)
+        else:
+            cmake = CMake(self)
+            cmake.install()
+            tools.rmdir(os.path.join(self.package_folder, "lib", "cmake"))
 
     def package_info(self):
         self.cpp_info.libs = tools.collect_libs(self)
+
+        if self.options.with_device_headers:
+            self.cpp_info.components["dnnLibrary"].set_property("cmake_target_name", "dnnLibrary::dnnLibrary")
+            self.cpp_info.components["dnnLibrary"].includedirs = ["include"]
+            if not self.options.header_only:
+                self.cpp_info.components["dnnLibrary"].libdirs = ["lib"]
+                self.cpp_info.components["dnnLibrary"].libs = ["dnn_lib"]
+                self.cpp_info.components["dnnLibrary"].requires = ["et-common-libs::cm-umode"]
+
+        if self.options.with_host_headers:
+            self.cpp_info.components["dnn_lib_host_generated_headers"].set_property("cmake_target_name", "dnnLibrary::dnn_lib_host_generated_headers")
+            self.cpp_info.components["dnn_lib_host_generated_headers"].includedirs = [os.path.join("include", "dnn_lib", "host_headers")]
+            self.cpp_info.components["dnn_lib_host_generated_headers"].libs = []
+            self.cpp_info.components["dnn_lib_host_generated_headers"].requires = []
